+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add Urdu language support
 - Add Malay language support
 - Add Swedish language support
+- Add Hebrew language support
 
 ## v3.1.0 — 2026-08-19